refactor(tasks): simplify subagent activity tracking
