+++ src/agents/embedded-agent-live-edit-diff.ts
@@ -1,6 +1,7 @@
 import { parseStreamingJson } from "@openclaw/ai/internal/runtime";
 import { normalizeLowercaseStringOrEmpty } from "@openclaw/normalization-core/string-coerce";
-import { resolveFileMutationToolName, type FileMutationToolName } from "./tool-mutation-names.js";
+import { countStreamingFileMutationLines } from "./file-mutation-args.js";
+import { resolveFileMutationToolName } from "./tool-mutation-names.js";
 
 const LIVE_EDIT_DIFF_MIN_INTERVAL_MS = 250;
 const LIVE_EDIT_DIFF_MAX_PARTIAL_JSON_CHARS = 1024 * 1024;
@@ -19,63 +20,6 @@ type LiveEditDiffProgress = {
   name: string;
   diff: { added: number; removed: number };
 };
-
-function countNewlines(value: unknown): number {
-  if (typeof value !== "string") {
-    return 0;
-  }
-  let count = 0;
-  for (let index = value.indexOf("\n"); index >= 0; index = value.indexOf("\n", index + 1)) {
-    count += 1;
-  }
-  return count;
-}
-
-function countEditLines(args: Record<string, unknown>): { added: number; removed: number } {
-  const replacements = Array.isArray(args.edits) ? args.edits : [args];
-  let added = 0;
-  let removed = 0;
-  for (const replacement of replacements) {
-    if (!replacement || typeof replacement !== "object" || Array.isArray(replacement)) {
-      continue;
-    }
-    const record = replacement as Record<string, unknown>;
-    added += countNewlines(record.newText ?? record.new_string);
-    removed += countNewlines(record.oldText ?? record.old_string);
-  }
-  return { added, removed };
-}
-
-function countPatchLines(patch: unknown): { added: number; removed: number } {
-  if (typeof patch !== "string") {
-    return { added: 0, removed: 0 };
-  }
-  let added = 0;
-  let removed = 0;
-  let lineStart = 0;
-  for (let lineEnd = patch.indexOf("\n"); lineEnd >= 0; lineEnd = patch.indexOf("\n", lineStart)) {
-    if (patch[lineStart] === "+") {
-      added += 1;
-    } else if (patch[lineStart] === "-") {
-      removed += 1;
-    }
-    lineStart = lineEnd + 1;
-  }
-  return { added, removed };
-}
-
-function countLiveEditDiff(
-  kind: FileMutationToolName,
-  args: Record<string, unknown>,
-): { added: number; removed: number } {
-  if (kind === "write") {
-    return { added: countNewlines(args.content), removed: 0 };
-  }
-  if (kind === "edit") {
-    return countEditLines(args);
-  }
-  return countPatchLines(args.input ?? args.patch);
-}
 
 function readToolCallBlock(event: Record<string, unknown>): Record<string, unknown> | undefined {
   const contentIndex = event.contentIndex;
@@ -160,7 +104,7 @@ export function updateLiveEditDiffProgress(
   // Parsing is the expensive part. Rate-limit it before touching cumulative JSON
   // so fragmented large arguments cannot create quadratic work on the event path.
   progress.lastCheckedAtMs = now;
-  const counted = countLiveEditDiff(kind, parseStreamingJson(partialJson));
+  const counted = countStreamingFileMutationLines(kind, parseStreamingJson(partialJson));
   // Streaming parses are best effort. Never move a visible counter backwards if
   // an incomplete JSON boundary temporarily exposes less of the same arguments.
   progress.added = Math.max(progress.added, counted.added);
